+++ src/task/task.service.ts
@@ -1,10 +1,15 @@
 import { Injectable } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
-import { NotFoundException } from '@nestjs/common/exceptions';
+import {
+  NotFoundException,
+  BadRequestException,
+} from '@nestjs/common/exceptions';
+import { Types } from 'mongoose';
 
 import { Task } from './task.schema';
 import { User } from 'src/user/user.schema';
+import { Category } from 'src/category/category.schema';
 import { CreateTaskDto } from './dtos/create-task.dto';
 import { QueryTaskDto } from './dtos/query-task.dto';
 
@@ -15,6 +20,20 @@ interface queryParams {
   deadline?: object;
 }
 
+interface createdTaskDoc {
+  __v: string;
+  title: string;
+  description: string;
+  categories: Category[];
+  isCompleted: boolean;
+  dateOfCompletion: Date;
+  links: Array<string>;
+  deadline: Date;
+  userId: User;
+  createdAt: Date;
+  updatedAt: Date;
+}
+
 @Injectable()
 export class TaskService {
   constructor(
@@ -23,17 +42,32 @@ export class TaskService {
   ) {}
 
   async findOne(userId: string, id: string): Promise<Task> {
-    const foundTask = await this.taskModel.findOne({ _id: id, userId });
+    if (!Types.ObjectId.isValid(id))
+      throw new BadRequestException('Invalid ObjectId');
+
+    const foundTask = await this.taskModel
+      .findOne({ _id: id, userId })
+      .select(['-__v']);
     if (!foundTask) throw new NotFoundException('Task not found');
 
     return foundTask;
   }
 
-  find(userId: string, query: QueryTaskDto): Promise<Task[]> {
+  async find(
+    userId: string,
+    query: QueryTaskDto,
+  ): Promise<{ tasks: Task[]; currentPage: number; totalPages: number }> {
+    const {
+      page = 1,
+      limit = 10,
+      isCompleted = null,
+      categories = null,
+      deadline = null,
+    } = query;
+
     let queryParams: queryParams = {
       userId: userId,
     };
-    const { isCompleted = null, categories = null, deadline = null } = query;
 
     if (isCompleted) {
       queryParams.isCompleted = JSON.parse(isCompleted);
@@ -77,12 +111,25 @@ export class TaskService {
       }
     }
 
-    return this.taskModel.find(queryParams).populate('categories');
+    const count = await this.taskModel.countDocuments(queryParams);
+
+    const totalPages = Math.ceil(count / limit);
+
+    const foundTasks = await this.taskModel
+      .find(queryParams)
+      .populate('categories')
+      .limit(limit * 1)
+      .skip((page - 1) * limit)
+      .select(['-__v'])
+      .exec();
+
+    return { tasks: foundTasks, currentPage: page, totalPages };
   }
 
   async create(userId: string, createTaskDto: CreateTaskDto): Promise<Task> {
     const createdTask = await this.taskModel.create({
       userId,
+      dateOfCompletion: createTaskDto.isCompleted ? new Date() : null,
       ...createTaskDto,
     });
 
@@ -90,7 +137,12 @@ export class TaskService {
       $push: { tasks: createdTask._id },
     });
 
-    return createdTask;
+    await createdTask.populate('categories');
+
+    const { __v, ...createdTaskData } =
+      createdTask.toObject() as createdTaskDoc;
+
+    return createdTaskData;
   }
 
   async update(
@@ -98,23 +150,28 @@ export class TaskService {
     id: string,
     attrs: Partial<Task>,
   ): Promise<Task> {
+    if (!Types.ObjectId.isValid(id))
+      throw new BadRequestException('Invalid ObjectId');
+
     if (attrs.isCompleted === true) {
       attrs.dateOfCompletion = new Date();
     } else if (attrs.isCompleted === false) {
       attrs.dateOfCompletion = null;
     }
 
-    const updatedTask = await this.taskModel.findOneAndUpdate(
-      { _id: id, userId },
-      attrs,
-      { new: true },
-    );
+    const updatedTask = await this.taskModel
+      .findOneAndUpdate({ _id: id, userId }, attrs, { new: true })
+      .populate('categories')
+      .select(['-__v']);
     if (!updatedTask) throw new NotFoundException('Task not found');
 
     return updatedTask;
   }
 
   async remove(userId: string, id: string): Promise<Task> {
+    if (!Types.ObjectId.isValid(id))
+      throw new BadRequestException('Invalid ObjectId');
+
     const deletedTask = await this.taskModel.findOneAndDelete({
       _id: id,
       userId,
@@ -127,5 +184,56 @@ export class TaskService {
     }
 
     return;
+  }
+
+  async getStats(userId: string) {
+    const date = new Date();
+    const year = date.getFullYear();
+    const month =
+      date.getMonth() + 1 < 10
+        ? `0${date.getMonth() + 1}`
+        : date.getMonth() + 1;
+    const day = date.getDate();
+    const tomorrowMidnight = new Date(`${year}-${month}-${day + 1}`);
+
+    const today = new Date(`${year}-${month}-${day}`);
+    const foundTasks = await this.taskModel.find({
+      userId: userId,
+      dateOfCompletion: {
+        $lte: tomorrowMidnight,
+        $gte: new Date(today.setDate(today.getDate() - 10)),
+      },
+    });
+
+    const stats = [];
+
+    for (let i = 0; i < 10; i++) {
+      const now = new Date(`${year}-${month}-${day}`);
+      const now2 = new Date(`${year}-${month}-${day}`);
+      const now3 = new Date(`${year}-${month}-${day}`);
+
+      const dayStats = {
+        date: new Date(now.setDate(now.getDate() - i)),
+        counter: 0,
+      };
+
+      const dayBefore = new Date(now2.setDate(now2.getDate() - i));
+      const dayAfter = new Date(now3.setDate(now3.getDate() - i + 1));
+
+      foundTasks.forEach((task) => {
+        if (
+          task.dateOfCompletion >= dayBefore &&
+          task.dateOfCompletion < dayAfter
+        ) {
+          dayStats.counter++;
+        }
+      });
+
+      stats.push(dayStats);
+    }
+
+    stats.reverse();
+
+    return stats;
   }
 }
